refactor: centralize consent constants
