+++ src/lib/consent.ts
@@ -1,0 +1,11 @@
+export const COOKIE_KEY = "dorsium_consent";
+
+export type ConsentPrefs = {
+  analytics: boolean;
+  marketing: boolean;
+};
+
+export const defaultPrefs: ConsentPrefs = {
+  analytics: false,
+  marketing: false,
+};
